Water Surface
Added a water surface to the game. It's basically just another chunk higher up on the map using the same noise function, but uses the cosine to change the height over time and make it look more like waves. I also added transparency to my block shader so that you can see through the water, but it made every other texture transparent as well. Whoops. I'll deal with that later. The water also kills the frame rate which I will deal with later.

I removed a lot of unnecessary garbage and cleaned up some of the code that deals with block faces so that we can exclude creating some vertices that we'll never see.