Chunk Edges and Curved Noise
Made it so that faces on the edges of chunks don't render if they're hidden by a neighboring chunk.

Added an AnimationCurve component to our ChunkManager object so that we can more easily customize the terrain. The curve allows us to make it so there are more valleys, smooth out slopes, steepen slopes, etc etc. The curve is normalized so it can be easily applied to our noise (which is also normalized) and affect the height value in some way shape or form.

Split up the functions that initialize, generate, and render the chunk data so that we can do something in the future to speed it up. Right now, It just helps cut down time when checking faces on the edges because all the chunks around it will already have their data populated by the time we get to the rendering checks.

Renamed some functions and added comments for more clarity because I kept confusing myself.