Fish Update
Refactoring of functions that deal with the boids (fish). Added a FieldOfView class that makes it so the fish only see other fish in front of them. This helped with the stuttering issues, allows for more customization of the fish schooling/swarming, and smoothed out the rotations. Still needs a little tweaking, but I think I'm close to where I want it to be. Set up a basic framework for detecting collisions with the environment. It doesn't work yet, but it's there.

Also created a test scene for just the fish so that we don't have to deal with waiting 30 seconds for the terrain to generate every time we want to test changes to the fish.

Added new fish and models. Currently have clown fish, blue tangs, puffer fish, and jellyfish.

Working on a shader to animate the fish swimming, but based on how it's going, I might just make an animation instead.