Threading!
Finally got threads implemented for chunk generation. This has significantly sped up loading chunks from ~30-60 seconds for 144 chunks (sized 32x32x32) to ~0.5 seconds at most.
Only minor drawback is that due to the way I did the threading, the heightmap data for all the chunks are not loaded before the meshes render. So I have to figure out a new way to prevent rendering faces on the edges of chunks. At the moment, I'm still consistently getting 200+fps and not hitting Unity's mesh vertex limit, so this is not really a big concern. Once more things are happening on the screen, I'll need to optimize this even more, but we're not there yet.